[refactor] swagger: 구현체와 인터페이스 리팩토링
- 인터페이스 = API 명세/Swagger용
- 구현체 = 실제 Spring MVC 요청 처리(@RequestBody, @AuthenticationPrincipal 등)
