+++ src/main/java/org/example/ootoutfitoftoday/domain/dashboard/controller/UserDashboardControllerImpl.java
@@ -8,8 +8,10 @@ import org.example.ootoutfitoftoday.domain.auth.dto.AuthUser;
 import org.example.ootoutfitoftoday.domain.dashboard.exception.DashboardSuccessCode;
 import org.example.ootoutfitoftoday.domain.dashboard.service.query.user.DashboardUserQueryService;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.time.LocalDate;
@@ -24,7 +26,7 @@ public class UserDashboardControllerImpl implements UserDashboardController {
     @Override
     @GetMapping("/overview")
     public ResponseEntity<Response<DashboardUserSummaryResponse>> getUserDashboardSummary(
-            AuthUser authUser
+            @AuthenticationPrincipal AuthUser authUser
     ) {
 
         return Response.success(dashboardUserQueryService.getUserDashboardSummary(authUser.getUserId()), DashboardSuccessCode.DASHBOARD_USER_SUMMARY_OK);
@@ -33,8 +35,8 @@ public class UserDashboardControllerImpl implements UserDashboardController {
     @Override
     @GetMapping("/statistics")
     public ResponseEntity<Response<DashboardUserWearStatisticsResponse>> getUserWearStatistics(
-            AuthUser authUser,
-            LocalDate baseDate
+            @AuthenticationPrincipal AuthUser authUser,
+            @RequestParam(required = false) LocalDate baseDate
     ) {
 
         return Response.success(dashboardUserQueryService.getUserWearStatistics(authUser.getUserId(), baseDate), DashboardSuccessCode.DASHBOARD_USER_STATISTICS_OK);
