+++ src/main/java/org/example/ootoutfitoftoday/domain/donation/controller/DonationCenterControllerImpl.java
@@ -9,6 +9,7 @@ import org.example.ootoutfitoftoday.domain.donation.service.query.DonationCenter
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
@@ -24,10 +25,10 @@ public class DonationCenterControllerImpl implements DonationCenterController {
     @Override
     @GetMapping("/search")
     public ResponseEntity<Response<List<DonationCenterSearchResponse>>> searchNearbyDonationCenters(
-            Double latitude,
-            Double longitude,
-            Integer radius,
-            String keyword
+            @RequestParam Double latitude,
+            @RequestParam Double longitude,
+            @RequestParam(required = false) Integer radius,
+            @RequestParam(required = false) String keyword
     ) {
         log.info("주변 기부처 검색 요청 - 위도: {}, 경도: {}, 반경: {}m, 키워드: {}",
                 latitude, longitude, radius, keyword);
